third_party/WebKit/LayoutTests/fast/viewport/viewport-72-expected.txt

CONSOLE MESSAGE: line 3: The key "x;initial-scale" is not recognized and ignored.
CONSOLE MESSAGE: line 3: Error parsing a meta element's content: ';' is not a valid key-value pair separator. Please use ',' instead.
ALERT: viewport size 100x110 scale 3.2 with limits [3.2, 5] and userScalable true
